Add query parser support
This adds the concept of default middleware, which runs before any user
supplied middleware functions.

This adds queryParser as the first default middleware, which checks if
req.query is set and if not parsed req.url with url.parse and URLSearchParams
from Node's url module.

In this implementation, req.query is a URLSearchParams instance, so you
can use that standard interface to access query params on the request object.