Guard middleware functions from calling next multiple times
Since the next function passed to middleware functions is a regular
old function, it can be called by the calling code as many times
as it wants (or not at all).

This wraps all next functions passed to middlewares in a callOnce
higher order function, which guards against the function being called
more than once.